+++ common/ASC.Data.Storage/S3/S3Storage.cs
@@ -190,7 +190,7 @@ namespace ASC.Data.Storage.S3
         {
             var contentDisposition = string.Format("attachment; filename={0};",
                                                    HttpUtility.UrlPathEncode(attachmentFileName));
-            if (attachmentFileName.Any(c => (int)c >= 0 && (int)c <= 127))
+            if (attachmentFileName.Any(c => c >= 0 && c <= 127))
             {
                 contentDisposition = string.Format("attachment; filename*=utf-8''{0};",
                                                    HttpUtility.UrlPathEncode(attachmentFileName));
