+++ web/ASC.Web.Core/Files/ContentDispositionUtil.cs
@@ -36,7 +36,7 @@ namespace ASC.Web.Core.Files
         {
             // If fileName contains any Unicode characters, encode according
             // to RFC 2231 (with clarifications from RFC 5987)
-            if (fileName.Any(c => (int)c > 127))
+            if (fileName.Any(c => c > 127))
             {
                 //.netcore
                 var str = withoutBase
